+++ src/js/api.ts
@@ -268,6 +268,8 @@ export class PyodideAPI_ {
    *        Defaults to ``"<exec>"``. If a custom file name is given, the
    *        traceback for any exception that is thrown will show source lines
    *        (unless the given file name starts with ``<`` and ends with ``>``).
+   * @param options.dedent An optional boolean indicating whether ``code``
+   *        should be dedented before being run. Defaults to ``true``.
    * @returns The result of the Python code translated to JavaScript. See the
    *          documentation for :py:func:`~pyodide.code.eval_code` for more info.
    * @example
@@ -288,7 +290,12 @@ export class PyodideAPI_ {
    */
   static runPython(
     code: string,
-    options: { globals?: PyProxy; locals?: PyProxy; filename?: string } = {},
+    options: {
+      globals?: PyProxy;
+      locals?: PyProxy;
+      filename?: string;
+      dedent?: boolean;
+    } = {},
   ): any {
     if (!options.globals) {
       options.globals = API.globals;
@@ -333,11 +340,18 @@ export class PyodideAPI_ {
    *        Defaults to ``"<exec>"``. If a custom file name is given, the
    *        traceback for any exception that is thrown will show source lines
    *        (unless the given file name starts with ``<`` and ends with ``>``).
+   * @param options.dedent An optional boolean indicating whether ``code``
+   *        should be dedented before being run. Defaults to ``true``.
    * @returns The result of the Python code translated to JavaScript.
    */
   static async runPythonAsync(
     code: string,
-    options: { globals?: PyProxy; locals?: PyProxy; filename?: string } = {},
+    options: {
+      globals?: PyProxy;
+      locals?: PyProxy;
+      filename?: string;
+      dedent?: boolean;
+    } = {},
   ): Promise<any> {
     if (!options.globals) {
       options.globals = API.globals;